+++ core/templates/components/state-editor/state-editor.component.html
@@ -1,0 +1,112 @@
+<md-card class="oppia-editor-card-with-avatar">
+  <div class="state-content-header-container" ng-click="toggleConceptCard()">
+    <div class="state-content-header oppia-mobile-collapsible-card-header">
+      <h3 class="oppia-exp-content-card-header">Content</h3>
+      <i class="fa fa-caret-down"
+         ng-if="!conceptCardIsShown"
+         aria-hidden="true">
+      </i>
+      <i class="fa fa-caret-up"
+         ng-if="conceptCardIsShown"
+         aria-hidden="true">
+      </i>
+    </div>
+  </div>
+  <div class="oppia-mobile-collapsible-card-content">
+    <div class="oppia-editor-card-body oppia-editor-concept-container" ng-if="conceptCardIsShown">
+      <img ng-if="!windowIsNarrow" ng-src="<[oppiaBlackImgUrl]>" alt="" class="oppia-editor-card-avatar">
+      <div id="tutorialStateContent" class="oppia-editor-card-section state-content-editor-parent-container">
+        <state-content-editor on-save-state-content="$ctrl.onSaveStateContent"
+                              show-mark-all-audio-as-needing-update-modal-if-required="$ctrl.showMarkAllAudioAsNeedingUpdateModalIfRequired"
+                              state-content-placeholder="$ctrl.getStateContentPlaceholder()"
+                              state-content-save-button-placeholder="$ctrl.getStateContentSaveButtonPlaceholder()">
+        </state-content-editor>
+      </div>
+    </div>
+  </div>
+</md-card>
+
+<div ng-show="$ctrl.isInteractionShown()">
+  <state-interaction-editor on-save-state-content="$ctrl.onSaveStateContent"
+                            on-save-interaction-id="$ctrl.onSaveInteractionId"
+                            on-save-interaction-customization-args="$ctrl.onSaveInteractionCustomizationArgs"
+                            on-save-next-content-id-index="$ctrl.onSaveNextContentIdIndex"
+                            on-save-solution="$ctrl.onSaveSolution"
+                            recompute-graph="$ctrl.recomputeGraph"
+                            show-mark-all-audio-as-needing-update-modal-if-required="$ctrl.showMarkAllAudioAsNeedingUpdateModalIfRequired">
+  </state-interaction-editor>
+</div>
+
+<div ng-hide="!interactionIdIsSet || currentStateIsTerminal">
+  <state-responses add-state="$ctrl.addState"
+                   navigate-to-state="$ctrl.navigateToState"
+                   on-responses-initialized="reinitializeEditor"
+                   on-save-inapplicable-skill-misconception-ids="$ctrl.onSaveInapplicableSkillMisconceptionIds"
+                   on-save-interaction-answer-groups="$ctrl.onSaveInteractionAnswerGroups"
+                   on-save-interaction-default-outcome="$ctrl.onSaveInteractionDefaultOutcome"
+                   on-save-next-content-id-index="$ctrl.onSaveNextContentIdIndex"
+                   on-save-solicit-answer-details="$ctrl.onSaveSolicitAnswerDetails"
+                   refresh-warnings="$ctrl.refreshWarnings"
+                   show-mark-all-audio-as-needing-update-modal-if-required="$ctrl.showMarkAllAudioAsNeedingUpdateModalIfRequired">
+  </state-responses>
+  <div>
+    <state-hints-editor on-save-hints="$ctrl.onSaveHints"
+                        on-save-next-content-id-index="$ctrl.onSaveNextContentIdIndex"
+                        on-save-solution="$ctrl.onSaveSolution"
+                        show-mark-all-audio-as-needing-update-modal-if-required="$ctrl.showMarkAllAudioAsNeedingUpdateModalIfRequired">
+    </state-hints-editor>
+  </div>
+  <div ng-if="currentInteractionCanHaveSolution">
+    <state-solution-editor on-save-solution="$ctrl.onSaveSolution"
+                           refresh-warnings="$ctrl.refreshWarnings"
+                           show-mark-all-audio-as-needing-update-modal-if-required="$ctrl.showMarkAllAudioAsNeedingUpdateModalIfRequired">
+    </state-solution-editor>
+  </div>
+  <div ng-if="$ctrl.explorationIsLinkedToStory">
+    <state-skill-editor (on-save-state-content)="$ctrl.onSaveStateContent($event)"
+                        (on-save-linked-skill-id)="$ctrl.onSaveLinkedSkillId($event)">
+    </state-skill-editor>
+  </div>
+</div>
+<style>
+  state-editor .state-content-header-container {
+    display: none;
+    padding: 0 30px;
+  }
+  state-editor .state-content-header {
+    align-content: center;
+    align-items: center;
+    display: flex;
+    justify-content: space-between;
+    padding: 30px 0 15px;
+  }
+  state-editor .state-content-header i {
+    display: none;
+  }
+  state-editor .oppia-exp-content-card-header {
+    font-size: 18px;
+    margin: 0;
+  }
+  state-editor .state-solution-content-container {
+    padding: 15px 30px;
+  }
+  state-editor .oppia-add-hint-button-container {
+    margin: 5px 0 15px;
+  }
+  state-editor .state-content-editor-parent-container {
+    padding-bottom: 36px;
+    padding-top: 36px;
+  }
+  @media screen and (max-width: 768px) {
+    state-editor .state-content-header-container {
+      display: block;
+      padding: 0;
+    }
+    state-editor .state-content-header {
+      padding: 18px 15px;
+    }
+    state-editor .state-content-header i {
+      display: block;
+    }
+  }
+</style>
