+++ scripts/check_frontend_test_coverage.py
@@ -237,7 +237,6 @@ NOT_FULLY_COVERED_FILENAMES = [
     'solution-explanation-editor.directive.ts',
     'state-card.model.ts',
     'state-content-editor.directive.ts',
-    'state-editor.directive.ts',
     'state-hints-editor.directive.ts',
     'state-interaction-editor.directive.ts',
     'state-property.service.ts',
